feat/add modal component

+++ src/components/Cards/HotelCard/index.tsx
@@ -1,26 +1,25 @@
+import { useAuth } from "../../../contexts/AuthContext";
 import { useHotel } from "../../../contexts/HotelContext";
+import { iHotel } from "../../../interface";
 import { StyledHotelCard } from "./style";
 
-interface hotelProps {
-    id: string
-    name: string
-    city: string
-    img: string
+interface CardProps {
+  hotel: iHotel;
 }
 
-export const HotelCard = (hotel: hotelProps) => {
-const {deleteHotel, setHotel} = useHotel()
+export const HotelCard = ({hotel}: CardProps) => {
+  const { setShowModal } = useAuth();
+  const { deleteHotel } = useHotel();
   return (
     <>
-      <StyledHotelCard>
+      <StyledHotelCard onClick={()=>setShowModal(hotel.full_url)}>
         <figure>
-          <img src={hotel.img} alt="Cover" />
+          <img src={hotel.full_url} alt="Cover" />
         </figure>
         <h3>{hotel.name}</h3>
-        <h4>{hotel.city}</h4>
+        <h4>{hotel.address.city}</h4>
         <div>
-          {/* <button onClick={(e)=>setHotel(e.)}>editar</button> */}
-          <button onClick={()=>deleteHotel(hotel.id)}>deletar</button>
+          <button onClick={() => deleteHotel(hotel.id)}>deletar</button>
         </div>
       </StyledHotelCard>
     </>
